requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
distlib==0.3.7
Django==4.1.5
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-extensions==3.2.3
django-filter==22.1
djangorestframework==3.14.0
filelock==3.13.1
mysql-connector-python==8.1.0
mysqlclient==2.2.1
Pillow==10.1.0
platformdirs==4.0.0
protobuf==4.21.12
pytz==2024.1
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.24.7
